--- deploy/deploy_box.ts
@@ -1,13 +1,0 @@
-import { deployContract } from './utils';
-
-export default async function () {
-  await deployContract(
-    'BoxDemo',
-    [],
-    {
-      noVerify: false,
-      upgradable: true,
-    },
-    ['42'],
-  );
-}
