--- deploy/nft/verifyContract.ts
@@ -1,5 +1,0 @@
-import { verifyContractByName } from '../utils';
-
-export default async function () {
-  await verifyContractByName('', '', []);
-}
