+++ test/nft/NovaLynksNFT.test.ts
@@ -54,102 +54,6 @@ describe('NovaLynksNFT', function () {
     lynkAddr = await Lynk.getAddress();
   });
 
-  it('mint TradeMark error', async function () {
-    tradeAddr = await TradeMark.getAddress();
-    const domain = {
-      name: 'TradeMark',
-      version: '0',
-      chainId: 31337,
-      verifyingContract: tradeAddr,
-    };
-
-    let types = {
-      MintAuth: [
-        { name: 'to', type: 'address' },
-        { name: 'nonce', type: 'uint256' },
-        { name: 'tokenId', type: 'uint256' },
-        { name: 'amount', type: 'uint256' },
-        { name: 'expiry', type: 'uint256' },
-      ],
-    };
-
-    let message1 = {
-      to: addr1.address,
-      nonce: 1,
-      tokenId: 1,
-      amount: 1,
-      expiry: 1742630631000,
-    };
-    let message2 = {
-      to: addr1.address,
-      nonce: 1,
-      tokenId: 2,
-      amount: 1,
-      expiry: 1742630631000,
-    };
-    let message3 = {
-      to: addr1.address,
-      nonce: 1,
-      tokenId: 3,
-      amount: 1,
-      expiry: 1742630631000,
-    };
-    let message4 = {
-      to: addr1.address,
-      nonce: 1,
-      tokenId: 4,
-      amount: 1,
-      expiry: 1742630631000,
-    };
-
-    // mint TradeMark
-    signature = await owner.signTypedData(domain, types, message1);
-    await TradeMark['safeMint(address,uint256,uint256,uint256,uint256,bytes)'](
-      addr1.address,
-      1,
-      1,
-      1,
-      1742630631000,
-      signature,
-    );
-    signature = await owner.signTypedData(domain, types, message2);
-    await TradeMark['safeMint(address,uint256,uint256,uint256,uint256,bytes)'](
-      addr1.address,
-      1,
-      2,
-      1,
-      1742630631000,
-      signature,
-    );
-    signature = await owner.signTypedData(domain, types, message3);
-    await TradeMark['safeMint(address,uint256,uint256,uint256,uint256,bytes)'](
-      addr1.address,
-      1,
-      3,
-      1,
-      1742630631000,
-      signature,
-    );
-    signature = await owner.signTypedData(domain, types, message4);
-    await TradeMark['safeMint(address,uint256,uint256,uint256,uint256,bytes)'](
-      addr1.address,
-      1,
-      4,
-      1,
-      1742630631000,
-      signature,
-    );
-
-    let token1Bal = await TradeMark.balanceOf(addr1.address, 1);
-    let token2Bal = await TradeMark.balanceOf(addr1.address, 2);
-    let token3Bal = await TradeMark.balanceOf(addr1.address, 3);
-    let token4Bal = await TradeMark.balanceOf(addr1.address, 4);
-    expect(token1Bal).to.equal(1);
-    expect(token2Bal).to.equal(1);
-    expect(token3Bal).to.equal(1);
-    expect(token4Bal).to.equal(1);
-  });
-
   it('mint Lynk without SBT error', async function () {
     lynkAddr = await Lynk.getAddress();
 
@@ -230,7 +134,7 @@ describe('NovaLynksNFT', function () {
       signature,
     );
     let token1Bal = await TradeMark.balanceOf(addr1.address, 1);
-    expect(token1Bal).to.equal(3);
+    expect(token1Bal).to.equal(2);
   });
 
   it('mint Lynk error', async function () {
@@ -255,10 +159,10 @@ describe('NovaLynksNFT', function () {
     let token2Bal = await TradeMark.balanceOf(addr1.address, 2);
     let token3Bal = await TradeMark.balanceOf(addr1.address, 3);
     let token4Bal = await TradeMark.balanceOf(addr1.address, 4);
-    expect(token1Bal).to.equal(2);
-    expect(token2Bal).to.equal(2);
-    expect(token3Bal).to.equal(2);
-    expect(token4Bal).to.equal(2);
+    expect(token1Bal).to.equal(1);
+    expect(token2Bal).to.equal(1);
+    expect(token3Bal).to.equal(1);
+    expect(token4Bal).to.equal(1);
   });
 
   it('mint Lynk with Auth error', async function () {
@@ -287,26 +191,5 @@ describe('NovaLynksNFT', function () {
     const balance = await Lynk.balanceOf(addr1.address);
     console.log('Lynk balance Auth:', balance); // mint success
     expect(balance).to.equal(2);
-  });
-
-  it('1155 transfer success', async function () {
-    const addr1TradeMark = new ethers.Contract(await TradeMark.getAddress(), TradeMark.interface, addr1);
-    await addr1TradeMark.safeTransferFrom(addr1.address, addr2.address, 1, 1, '0x');
-    let addr2token1Bal = await TradeMark.balanceOf(addr2.address, 1);
-    let addr1token1Bal = await TradeMark.balanceOf(addr1.address, 1);
-    console.log('addr2token1Bal:', addr2token1Bal);
-    console.log('addr1token1Bal:', addr1token1Bal);
-    expect(addr2token1Bal).to.equal(1);
-    expect(addr1token1Bal).to.equal(1);
-  });
-
-  it('1155 burn success', async function () {
-    const addr1TradeMark = new ethers.Contract(await TradeMark.getAddress(), TradeMark.interface, addr1);
-    const addr2TradeMark = new ethers.Contract(await TradeMark.getAddress(), TradeMark.interface, addr2);
-    addr1TradeMark.setApprovalForAll(addr2.address, true);
-    await expect(addr2TradeMark.burn(addr1.address, 1, 2)).to.be.revertedWith('ERC1155: burn amount exceeds balance');
-    await addr2TradeMark.burn(addr1.address, 1, 1);
-    let token1Bal = await TradeMark.balanceOf(addr1.address, 1);
-    expect(token1Bal).to.equal(0);
   });
 });
